Simplify tests, moving them all into one file
